last step

+++ lib/presentation/widgets/single_message_page_widgets/delete_button.dart
@@ -4,53 +4,27 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:lostsapp/logic/cubit/auth_cubit.dart';
 import 'package:lostsapp/logic/cubit/delete_message_cubit.dart';
 
+import '../delete_confirm_dialog.dart';
+
 class DeleteButton extends StatelessWidget {
   final String messageId;
 
   const DeleteButton({Key? key, required this.messageId}) : super(key: key);
-  Future<bool> _showConfirmDialog<bool>(BuildContext context) async {
-    return await showDialog(
-        context: context,
-        builder: (context) {
-          return AlertDialog(
-            title: Text(
-              "Are you sure?",
-              style: TextStyle(
-                  fontWeight: FontWeight.bold,
-                  color: Theme.of(context).colorScheme.secondary),
-            ),
-            content: Text(
-              "Deleted Items can't be resotred",
-              style: TextStyle(color: Theme.of(context).colorScheme.secondary),
-            ),
-            actions: [
-              TextButton(
-                child: const Text("Cancel"),
-                onPressed: () => Navigator.of(context).pop(false),
-              ),
-              TextButton(
-                child: const Text(
-                  "Delete",
-                  style: TextStyle(color: Colors.red),
-                ),
-                onPressed: () => Navigator.of(context).pop(true),
-              )
-            ],
-          );
-        });
-  }
 
   @override
   Widget build(BuildContext context) {
+    final pageWidth = MediaQuery.of(context).size.width;
     return Container(
-      padding: const EdgeInsets.only(left: 10, right: 10, top: 10, bottom: 25),
+      padding: pageWidth < 750
+          ? const EdgeInsets.only(left: 20, right: 20, top: 10, bottom: 25)
+          : EdgeInsets.symmetric(horizontal: pageWidth / 4, vertical: 15),
       child: ElevatedButton(
         style: ButtonStyle(
           backgroundColor: MaterialStateProperty.all<Color>(Colors.red),
         ),
         onPressed: () async {
-          bool toDelete = await _showConfirmDialog(context);
-          if (toDelete) {
+          bool? toDelete = await showDeleteConfirmDialog(context, "messages");
+          if (toDelete != null && toDelete) {
             await context.read<DeleteMessageCubit>().deleteMessage(
                 context.read<AuthCubit>().user!.token, messageId);
           }
